Add report...
I heard that the department would upload it somewhere(where?) for the purpose that anyone could look up (I hope. So dropping it here too is ok, I think. It embarassing somehow to show a work that disappointing for me somehow. well, if anyone take a look and get some good idea or inspiration from it, i would be proud of you. LOL